Setup sbt as a separate step
Calling `sbt --version` twice appears redundant, but it accomplishes two things.  The first call works around an issue in which the first call to `sbt` does nothing on Ubuntu.  This is a known issue.  The first call just lets it do nothing and sets up the second call.  The second call initiates the process of downloading, configuring, and installing `sbt`.  The value of the version isn't important, just that we're kicking off the process to actually set up `sbt`.  That results in a lot of log spew, so we've moved it to a separate step.  This will allow us to highlight the results of actually running the test in its own phase.